Fixing initial text being set without accessing textTransaction object publicly
